Create a file refactor.rb and move class inside that file
